Початкова версія сторінки клієнтів на основі старої бд з призми сутностями

+++ backend/src/main/java/com/aksi/dto/client/ClientDTO.java
@@ -3,6 +3,7 @@ package com.aksi.dto.client;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import java.math.BigDecimal;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
@@ -70,7 +71,7 @@ public class ClientDTO {
     /**
      * Загальна сума замовлень
      */
-    private Double totalSpent;
+    private BigDecimal totalSpent;
     
     /**
      * Кількість замовлень
